# Conflicts:
#	src/Mpociot/CaptainHook/CaptainHookServiceProvider.php

+++ src/Mpociot/CaptainHook/Jobs/TriggerWebhooksJob.php
@@ -6,6 +6,7 @@ use GuzzleHttp\Client;
 use GuzzleHttp\Middleware;
 use GuzzleHttp\Promise\PromiseInterface;
 use Illuminate\Bus\Queueable;
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Str;
 use Mpociot\CaptainHook\WebhookLog;
 use Illuminate\Queue\SerializesModels;
@@ -40,6 +41,13 @@ class TriggerWebhooksJob implements ShouldQueue
      */
     public function __construct($webhooks, $eventData)
     {
+        $model = end($eventData);
+
+        if ($model instanceof Model && method_exists($model, 'getWebhooksQueue')) {
+            $this->queue = $model->getWebhooksQueue();
+        } else {
+            $this->queue = 'default';
+        }
         $this->eventData = $eventData;
         $this->webhooks = $webhooks;
     }
@@ -81,52 +89,56 @@ class TriggerWebhooksJob implements ShouldQueue
         }
 
         $logging = $config->get('captain_hook.log.active');
+        $filterWebhook = $this->resolveCallable($config->get('captain_hook.filter_webhook'), 'filter');
         $transformer = $this->resolveCallable($config->get('captain_hook.transformer'), 'transform');
         $responseCallback = $this->resolveCallable($config->get('captain_hook.response_callback'), 'handle');
 
         foreach ($this->webhooks as $webhook) {
-            if ($logging) {
-                if ($config->get('captain_hook.log.storage_quantity') != -1 &&
-                    $webhook->logs()->count() >= $config->get('captain_hook.log.storage_quantity')) {
-                    $webhook->logs()->orderBy('updated_at', 'asc')->first()->delete();
-                }
-                $log = new WebhookLog([
-                    'webhook_id' => $webhook['id'],
-                    'url' => $webhook['url'],
-                ]);
-                $middleware = Middleware::tap(function (RequestInterface $request, $options) use ($log) {
-                    $log->payload_format = isset($request->getHeader('Content-Type')[0]) ? $request->getHeader('Content-Type')[0] : null;
-                    $log->payload = $request->getBody()->getContents();
-                }, function ($request, $options, PromiseInterface $response) use ($log, $webhook, $responseCallback) {
-                    $response->then(function (ResponseInterface $response) use ($log, $webhook, $responseCallback) {
-                        $log->status = $response->getStatusCode();
-                        $log->response = $response->getBody()->getContents();
-                        $log->response_format = $log->payload_format = isset($response->getHeader('Content-Type')[0]) ? $response->getHeader('Content-Type')[0] : null;
+            if ($filterWebhook($this->eventData, $webhook)) {
+                if ($logging) {
+                    if ($config->get('captain_hook.log.storage_quantity') != -1 &&
+                        $webhook->logs()->count() >= $config->get('captain_hook.log.storage_quantity')) {
+                        $webhook->logs()->orderBy('updated_at', 'asc')->first()->delete();
+                    }
+                    $log = new WebhookLog([
+                        'webhook_id' => $webhook['id'],
+                        'url' => $webhook['url'],
+                    ]);
+                    $middleware = Middleware::tap(function (RequestInterface $request, $options) use ($log) {
+                        $log->payload_format = isset($request->getHeader('Content-Type')[0]) ? $request->getHeader('Content-Type')[0] : null;
+                        $log->payload = $request->getBody()->getContents();
+                    }, function ($request, $options, PromiseInterface $response) use ($log, $webhook, $responseCallback) {
+                        $response->then(function (ResponseInterface $response) use ($log, $webhook, $responseCallback) {
+                            $log->status = $response->getStatusCode();
+                            $log->response = $response->getBody()->getContents();
+                            $log->response_format = $log->payload_format = isset($response->getHeader('Content-Type')[0]) ? $response->getHeader('Content-Type')[0] : null;
 
-                        $log->save();
+                            $log->save();
 
-                        // Retry this job if the webhook response didn't give us a HTTP 200 OK
-                        if ($response->getStatusCode() >= 300 || $response->getStatusCode() < 200) {
-                            $this->release(30);
-                        }
+                            // Retry this job if the webhook response didn't give us a HTTP 200 OK
+                            if ($response->getStatusCode() >= 300 || $response->getStatusCode() < 200) {
+                                $this->release(30);
+                            }
 
-                        $responseCallback($webhook, $response);
+                            $responseCallback($webhook, $response);
+                        });
                     });
-                });
 
-                $client->post($webhook['url'], [
-                    'exceptions' => false,
-                    'body' => $transformer($this->eventData, $webhook),
-                    'verify' => false,
-                    'handler' => $middleware($client->getConfig('handler')),
-                ]);
-            } else {
-                $client->post($webhook['url'], [
-                    'exceptions' => false,
-                    'body' => $transformer($this->eventData, $webhook),
-                    'verify' => false,
-                    'timeout' => 10,
-                ]);
+                    $client->post($webhook['url'], [
+                        'exceptions' => false,
+                        'json' => $transformer($this->eventData, $webhook),
+                        'verify' => false,
+                        'handler' => $middleware($client->getConfig('handler')),
+                        'timeout' => 10
+                    ]);
+                } else {
+                    $client->post($webhook['url'], [
+                        'exceptions' => false,
+                        'json' => $transformer($this->eventData, $webhook),
+                        'verify' => false,
+                        'timeout' => 10,
+                    ]);
+                }
             }
         }
     }
